Created migration for os_volume attribute, added validation to CreateRequest, added attribute to Volume Model, return value in VolumeResource, added value to VolumeController's store() method, and tweaked the PrepareOsDisk job to set the flag to true when creating a volume.
#851 Introduce volume type property

+++ app/Http/Controllers/V2/VolumeController.php
@@ -93,6 +93,7 @@ class VolumeController extends BaseController
             'availability_zone_id',
             'capacity',
             'iops',
+            'os_volume',
         ]));
 
         $model->save();
@@ -143,6 +144,17 @@ class VolumeController extends BaseController
     {
         $volume = Volume::forUser(Auth::user())->findOrFail($volumeId);
         $instance = Instance::forUser(Auth::user())->findOrFail($request->get('instance_id'));
+
+        if ($instance->os_volume) {
+            return response()->json([
+                'errors' => [
+                    'title' => 'Detachment Failed',
+                    'detail' => 'The specified os volume cannot be detached.',
+                    'status' => 412,
+                    'source' => 'os_volume'
+                ]
+            ], 412);
+        }
 
         $volume->withSyncLock(function ($volume) use ($instance) {
             $instance->volumes()->detach($volume);
